requirements.txt
appdirs==1.4.3
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.2
Automat==0.6.0
channels==1.1.6
constantly==15.1.0
daphne==1.3.0
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.32.0
django-filter==1.1.0
django-model-utils==3.0.0
django-rest-auth==0.9.1
django-tables2==1.6.1
djangorestframework==3.7.7
hyperlink==17.2.1
incremental==17.5.0
Markdown==2.6.8
msgpack-python==0.4.8
oauthlib==2.0.2
packaging==16.8
psycopg2-binary==2.7.4
pyparsing==2.2.0
pyTelegramBotAPI==2.2.3
python3-openid==3.1.0
pytz==2018.3
redis==2.10.5
requests==2.7.0
requests-oauthlib==0.8.0
six==1.10.0
telebot==0.0.3
Twisted==17.9.0
txaio==2.8.0
zope.interface==4.4.2
